Struct secrets::boxed::Box

pub(crate) struct Box<T: Bytes> {
    ptr: NonNull<T>,
    len: usize,
    prot: Cell<Prot>,
    refs: Cell<u8>,
}

NOTE: This implementation is not meant to be exposed directly to end-users, and user-facing wrappers must be written with care to ensure they statically enforce the required invariants. These invariants are asserted with debug_assert in order to catch bugs at development-time, but these assertions will be compiled out in release-mode due to the expectation that they are enforced statically.

TODO: document invariants

Fields
ptr: NonNull<T>

the non-null pointer to the underlying protected memory

len: usize

the number of elements of T that can be stored in ptr

prot: Cell<Prot>

the pointer's current protection level

refs: Cell<u8>

the number of outstanding borrows; mutable borrows are tracked here even though there is a max of one, so that asserts can ensure invariants are obeyed

Methods

impl<T: Bytes> Box<T>

pub(crate) fn new_one<F>(init: F) -> Self where
F: FnOnce(&mut T),

Instantiates a new Box that can hold exactly one element of type T. The callback F will be used for initialization and will be called with a mutable reference to the element. The memory will be pre-filled with fixed bytes of arbitrary value.

pub(crate) fn new<F>(len: usize, init: F) -> Self where
F: FnOnce(&mut [T]),

Instantiates a new Box that can hold len elements of type T. The callback F will be used for initialization and will be called with a mutable reference to a slice containing these elements. The memory will be pre-filled with fixed bytes of arbitrary value.

pub(crate) fn len(&self) -> usize

Returns the number of elements in the Box.

pub(crate) fn is_empty(&self) -> bool

Returns true if the Box is empty.

pub(crate) fn size(&self) -> usize

Returns the size in bytes of the data contained in the Box. This does not include incidental metadata used in the implementation of Box itself, only the size of the data allocated on behalf of the user.

It is the maximum number of bytes that can be read from the internal pointer.

pub(crate) fn unlock(&self) -> &Self

Allows the contents of the Box to be read from. Any call to this function must be balanced with a call to lock. Mirroring Rust's borrowing rules, there may be any number of outstanding immutable unlocks (technically, limited by the max value of RefCount) or one mutable unlock.

pub(crate) fn unlock_mut(&mut self) -> &mut Self

Allows the contents of the Box to be read from and written to. Any call to this function must be balanced with a call to lock. Mirroring Rust's borrowing rules, there may be any number of outstanding immutable unlocks (technically, limited by the max value of RefCount) or one mutable unlock.

pub(crate) fn lock(&self)

Disables all access to the underlying memory. Must only be called to precisely balance prior calls to unlock and unlock_mut.

Calling this method in excess of the number of outstanding unlocks will result in a runtime panic. Omitting a call to this method and leaving an outstanding unlock will result in a runtime panic when this object is dropped.

pub(crate) unsafe fn as_ref(&self) -> &T

Converts the Box's contents into a reference. This must only happen while it is unlocked, and the reference must go out of scope before it is locked.

pub(crate) unsafe fn as_mut(&mut self) -> &mut T

Converts the Box's contents into a mutable reference. This must only happen while it is mutably unlocked, and the slice must go out of scope before it is locked.

pub(crate) fn as_slice(&self) -> &[T]

Converts the Box's contents into a slice. This must only happen while it is unlocked, and the slice must go out of scope before it is locked.

pub(crate) fn as_mut_slice(&mut self) -> &mut [T]

Converts the Box's contents into a mutable slice. This must only happen while it is mutably unlocked, and the slice must go out of scope before it is locked.

fn new_unlocked(len: usize) -> Self

Instantiates a new Box that can hold len elements of type T. This Box will be unlocked and must be locked before it is dropped.

TODO: make len a NonZero when it's stabilized and remove the related panic.

fn retain(&self, prot: Prot)

Performs the underlying retain half of the retain/release logic for monitoring outstanding calls to unlock.

fn release(&self)

Removes one outsdanding retain, and changes the memory protection level back to Prot::NoAccess when the number of outstanding retains reaches zero.

fn is_locked(&self) -> bool

Returns true if the protection level is [NoAccess]. Ignores ref count.

impl<T: Bytes + Randomizable> Box<T>

pub(crate) fn random(len: usize) -> Self

Instantiates a new Box with crypotgraphically-randomized contents.

impl<T: Bytes + Zeroable> Box<T>

pub(crate) fn zero(len: usize) -> Self

Instantiates a new Box whose backing memory is zeroed.