secrets::traits
Trait ConstantEq

pub trait ConstantEq: AsContiguousBytes {
    // Provided method
    fn constant_eq(&self, rhs: &Self) -> bool { ... }
}

A marker trait for types that can be compared for equality bitwise in constant time.

Note that this trait does not force types to be compared in constant time. When dealing with types that must be guaranteed to compare in constant time, it is highly encouraged to use wrappers that implement PartialEq by calling constant_eq.

Provided Methods

fn constant_eq(&self, rhs: &Self) -> bool

Compares self and rhs. Guaranteed to return false when the two arguments differen in size, and guaranteed to perform the bitwise comparison in constant O(size) time without short-circuiting.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.