secrets::traits
Trait Randomizable

pub unsafe trait Randomizable: AsContiguousBytes {
    // Provided method
    fn randomize(&mut self) { ... }
}

Types that can be safely initialized by setting their memory to a random value.

The trait is marked unsafe in order to restrict implementors to types that can safely have their underlying memory randomized.

Safety

This trait allows for overwriting a type’s memory with a cryptographically random bit pattern. If it is not legal to represent your type as any potential series of bits, then your type may not implement this trait.

Provided Methods

fn randomize(&mut self)

Randomizes the contents of self.