Module secrets::secret_vec

Container for SecretVec.

Structs

ItemMut

A mutable wrapper around an item in a SecretVec.

ItemRef

An immutable wrapper around an item in a SecretVec.

IterMut

An iterator for SecretVec that returns ItemMut.

IterRef

An iterator for SecretVec that returns ItemRef.

Ref

An immutable wrapper around the internal contents of a SecretVec. This wrapper Derefs to its slice representation for convenience.

RefMut

A mutable wrapper around the internal contents of a SecretVec. This wrapper Derefs to its slice representation for convenience.

SecretVec

A type for protecting variable-length secrets allocated on the heap.