Module secrets::traits

Marker traits to allow types to be contained as secrets.

Traits

AsContiguousBytes
Marker trait for types who are intrepretable as a series of contiguous bytes, where the exact size may not be known at compile-time. Any type that implements AsContiguousBytes must not exhibit undefined behavior when its underlying bits are set to any arbitrary bit pattern.

Bytes
A marker trait for types whose size is known at compile time and can be treated as raw buckets of bytes. Any type that implements Bytes must not exhibit undefined behavior when its underlying bits are set to any arbitrary bit pattern.

ConstantEq
A marker trait for types that can be compared for equality bitwise in constant time.

Randomizable
Types that can be safely initialized by setting their memory to a random value.

Zeroable
Types that can be safely initialized by setting their memory to all zeroes.