Trait secrets::traits::AsContiguousBytes

pub unsafe trait AsContiguousBytes {
    // Required methods
    fn size(&self) -> usize;
    fn as_u8_ptr(&self) -> *const u8;
    fn as_mut_u8_ptr(&mut self) -> *mut u8;

    // Provided methods
    fn as_bytes(&self) -> &[u8]  { ... }
    fn as_mut_bytes(&mut self) -> &mut [u8]  { ... }
}

Marker trait for types who are intrepretable as a series of contiguous bytes, where the exact size may not be known at compile-time. Any type that implements AsContiguousBytes must not exhibit undefined behavior when its underlying bits are set to any arbitrary bit pattern.

Safety

This trait must only be implemented for types that are interpreble as size bytes contiguous in memory.

Required Methods

fn size(&self) -> usize

Returns the size in bytes of Self.

fn as_u8_ptr(&self) -> *const u8

Returns a *const u8 pointer to the beginning of the data.

fn as_mut_u8_ptr(&mut self) -> *mut u8

Returns a *mut u8 pointer to the beginning of the data.

Provided Methods

fn as_bytes(&self) -> &[u8]

Returns a byte slice to the underlying data.

fn as_mut_bytes(&mut self) -> &mut [u8]

Returns a mutable byte slice to the underlying data.